requirements.txt
anyio==3.7.1
asgiref==3.7.2
certifi==2023.5.7
Django==4.1.10
django-cors-headers==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.0
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
Pillow==10.0.0
psycopg2==2.9.6
pymongo==3.12.3
pytz==2023.3
sniffio==1.3.0
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
